@@ -13,17 +13,33 @@
  * limitations under the License.
  */
 
-apply plugin : 'com.android.application'
+plugins {
+  id 'com.android.application'
+  id 'org.jetbrains.kotlin.android'
+}
 
 android {
-  compileSdkVersion 28
+  namespace 'com.android.grafika'
+  compileSdk 33
+
   defaultConfig {
     applicationId "com.google.grafika"
-    minSdkVersion 14
-    targetSdkVersion 28
+    minSdk 21
+    targetSdk 33
   }
+
+  compileOptions {
+    sourceCompatibility JavaVersion.VERSION_1_8
+    targetCompatibility JavaVersion.VERSION_1_8
+  }
+
+  kotlinOptions {
+    jvmTarget = '1.8'
+  }
+
+
 }
 
 dependencies {
-  implementation 'com.android.support:appcompat-v7:28.0.0'
+  implementation 'androidx.appcompat:appcompat:1.6.1'
 }
